chore(cfg) Moved config to root lib folder for more clarity
