+++ src/index.ts
@@ -2,4 +2,4 @@ export { initFederation } from './lib/init-federation';
 
 export * from './lib/handlers';
 export * from './lib/steps';
-export * from './lib/utils';
+export * from './lib/config';
