+++ src/lib/config/index.ts
@@ -1,0 +1,2 @@
+export { Config } from './config.contract';
+export { defaultConfig } from './config';
