--- src/lib/utils/index.ts
@@ -1,3 +1,0 @@
-
-export { Config } from './config/config.contract';
-export { defaultConfig } from './config/config';
